Create Kruskal's Algorithm
Kruskal's Algorithm in Java Language
Kruskal's Algorithm is used to find Minimum Spanning Tree.
A minimum spanning tree (MST) or minimum weight spanning tree for a weighted, connected, undirected graph is a spanning tree with a weight less than or equal to the weight of every other spanning tree. The weight of a spanning tree is the sum of weights given to each edge of the spanning tree.
A minimum spanning tree has (V – 1) edges where V is the number of vertices in the given graph.